+++ src/main.js
@@ -4,22 +4,29 @@ import createInstance from './instance/instance';
 // Import farmOS-map CSS.
 import './styles.css';
 
-// Define window.farmOS if it isn't already.
-if (typeof window.farmOS === 'undefined') {
-  window.farmOS = {};
-}
-
-// Add a farmOS.map object that is available globaly.
-window.farmOS.map = {
+// Export the default farmOS-map object - normally exposed globally at farmOS.map
+export default {
 
   // Initialize an array of farmOS map instances.
   instances: [],
 
-  // Initialize map behaviors.
+  // Behaviors which can be added on-demand by name via `instance.addBehavior`
   // These are objects with an attach() method that will be called when a map is
   // created. They receive the instance object so that they can perform
   // modifications to the map using instance methods.
-  behaviors: {},
+  namedBehaviors: {
+    rememberLayer: lazyLoadedBehavior('rememberLayer'),
+    edit: lazyLoadedBehavior('edit'),
+    google: lazyLoadedBehavior('google'),
+    measure: lazyLoadedBehavior('measure'),
+    snappingGrid: lazyLoadedBehavior('snappingGrid'),
+  },
+
+  // Map behaviors which will be automatically attached to all created map
+  // instances.
+  behaviors: {
+    rememberLayer: lazyLoadedBehavior('rememberLayer'),
+  },
 
   // Create a new farmOS map attached to a target element ID and add it to the
   // global instances array.
@@ -62,3 +69,14 @@ window.farmOS.map = {
     return this.instances.findIndex(instance => instance.target === target);
   },
 };
+
+function lazyLoadedBehavior(name) {
+  return {
+    async attach(instance, options) {
+      return import(/* webpackChunkName: "farmOS-map-[request]" */ `./behavior/${name}`).then((module) => {
+        const behavior = module.default;
+        behavior.attach(instance, options);
+      });
+    },
+  };
+}